feat: Update database driver to MySQL
The code changes in this commit update the database driver from PostgreSQL to MySQL. This change is necessary to connect to the MySQL database specified in the docker-compose.yml file. The new driver is imported from `gorm.io/driver/mysql` and the connection URL is modified accordingly. This commit also updates the docker-compose.yml file to use the `mariadb:10.5` image and sets the necessary environment variables for MySQL. The healthcheck command is also modified to use a custom script for checking the database's health.